# Conflicts:
#	TowerDefence/Assets/Magic Circle Fx.meta
#	TowerDefence/Assets/Magic Circle Fx/Resources.meta
#	TowerDefence/Assets/Magic Circle Fx/Resources/Prefabs.meta
#	TowerDefence/Assets/Magic Circle Fx/VFX_Materials.meta
#	TowerDefence/Assets/Plugins/Febucci/Text Animator/Extra/Typewriter Sound - Package.unitypackage.meta
#	TowerDefence/Assets/Technical Station/Models/Materials/No Name 2.mat.meta

+++ TowerDefence/Assets/_Project Main/Scripts/Unit/AttackUnitProperty.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 public class AttackUnitProperty : MonoBehaviour
 {
     public int attackRange; // 공격범위
-    public int damage; // 데미지(공격력)
+    public float damage; // 데미지(공격력)
     public int attackCount; // 공격대상
     public float speed; // 공격속도
     public int rotateSpeed; // 회전속도
